fix: make chained message regeneratable
